+++ changelog.md
@@ -1,6 +1,14 @@
 # Change log
+## 0.2.13
+_Released at 06-11-2025_
+
+### Summary
+- Autocreate now supports algorithms with parameters. 
+- Delete button is now disabled while a scenario is queued or running.
+- CQM-loader is now a configurable option
+
 ## 0.2.12
-_Released at 05-11-2025_
+_Released at 06-11-2025_
 
 ### Summary
 - Bug fixes